requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
openai==1.12.0
requests==2.31.0
werkzeug==3.0.1
